test: write and test new examples with noop and replace macros
